Fix all compiler warnings in MRPD and test code
This commit eliminates all C4244, C4267, C4312, C4311, C4047, C4101, and C4200
warnings that were causing potential runtime issues:

Core MRPD fixes:
- mmrp.c: Added explicit (int) casts for pointer arithmetic assignments
- msrp.c: Fixed pointer arithmetic and mrpd_send return value assignments
- mvrp.c: Added explicit (int) casts for pointer arithmetic assignments
- mrpw.c: Added (int) cast for pcap_sendpacket len parameter

Test framework fixes:
- mrp_doubles.c: Fixed HTIMER/int casting issues using intptr_t
- msrp_tests.cpp: Fixed size_t to unsigned int assignments
- msrp_pruning_tests.cpp: Added (int) casts for strlen return values
- msrp_pdu_tests.cpp: Removed unused evtid variable
- mvrp_pdu_tests.cpp: Removed unused evtid variable

Build system fixes:
- Updated cpputest CMakeLists.txt minimum version to 3.5 for modern CMake compatibility

All changes preserve functionality while eliminating potential data loss
warnings on 64-bit Windows builds. The CI pipeline now builds completely
warning-free.

+++ daemons/mrpd/mmrp.c
@@ -1030,7 +1030,7 @@ mmrp_emit_svcvectors(unsigned char *msgbuf, unsigned char *msgbuf_eof,
 	*mrpdu_msg_ptr = 0;
 	mrpdu_msg_ptr++;
 
-	*bytes_used = (mrpdu_msg_ptr - msgbuf);
+	*bytes_used = (int)(mrpdu_msg_ptr - msgbuf);
 	return 0;
  oops:
 	/* an internal error - caller should assume TXLAF */
@@ -1260,7 +1260,7 @@ mmrp_emit_macvectors(unsigned char *msgbuf, unsigned char *msgbuf_eof,
 	*mrpdu_msg_ptr = 0;
 	mrpdu_msg_ptr++;
 
-	*bytes_used = (mrpdu_msg_ptr - msgbuf);
+	*bytes_used = (int)(mrpdu_msg_ptr - msgbuf);
 	return 0;
  oops:
 	/* an internal error - caller should assume TXLAF */
@@ -1381,9 +1381,9 @@ int mmrp_txpdu(void)
 	} else
 		goto out;
 
-	msgbuf_len = mrpdu_msg_ptr - msgbuf;
+	msgbuf_len = (int)(mrpdu_msg_ptr - msgbuf);
 
-	bytes = mrpd_send(mmrp_socket, msgbuf, msgbuf_len, 0);
+	bytes = (int)mrpd_send(mmrp_socket, msgbuf, msgbuf_len, 0);
 #if LOG_MMRP
 	mrpd_log_printf("MMRP send PDU\n");
 #endif
